Issue-146: Renamed EntertainmentMapper -> CoreMapper.
