+++ app/src/main/java/com/kirchhoff/movies/mapper/core/ICoreMapper.kt
@@ -1,0 +1,11 @@
+package com.kirchhoff.movies.mapper.core
+
+import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
+import com.kirchhoff.movies.data.network.core.NetworkObjectWithName
+import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
+import com.kirchhoff.movies.data.ui.core.UIGenre
+
+interface ICoreMapper {
+    fun createUIEntertainmentCredits(credits: NetworkEntertainmentCredits): UIEntertainmentCredits
+    fun createUIGenre(item: NetworkObjectWithName): UIGenre
+}
