+++ app/src/main/java/com/kirchhoff/movies/mapper/movie/MovieDetailsMapper.kt
@@ -6,16 +6,15 @@ import com.kirchhoff.movies.data.network.details.movie.NetworkMovieDetails
 import com.kirchhoff.movies.data.network.details.movie.NetworkTrailer
 import com.kirchhoff.movies.data.network.details.movie.NetworkTrailersList
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
-import com.kirchhoff.movies.data.ui.core.UIGenre
 import com.kirchhoff.movies.data.ui.details.movie.UICountry
 import com.kirchhoff.movies.data.ui.details.movie.UIMovieDetails
 import com.kirchhoff.movies.data.ui.details.movie.UITrailer
 import com.kirchhoff.movies.data.ui.details.movie.UITrailersList
 import com.kirchhoff.movies.mapper.BaseMapper
-import com.kirchhoff.movies.mapper.entertainment.IEntertainmentMapper
+import com.kirchhoff.movies.mapper.core.ICoreMapper
 import com.kirchhoff.movies.repository.Result
 
-class MovieDetailsMapper(private val entertainmentMapper: IEntertainmentMapper) : BaseMapper(), IMovieDetailsMapper {
+class MovieDetailsMapper(private val coreMapper: ICoreMapper) : BaseMapper(), IMovieDetailsMapper {
     override fun createUIMovieDetails(movieDetailsResult: Result<NetworkMovieDetails>): Result<UIMovieDetails> =
         when (movieDetailsResult) {
             is Result.Success -> Result.Success(createUIMovieDetails(movieDetailsResult.data))
@@ -24,7 +23,7 @@ class MovieDetailsMapper(private val entertainmentMapper: IEntertainmentMapper) 
 
     override fun createUIEntertainmentCredits(movieCreditsResult: Result<NetworkEntertainmentCredits>): Result<UIEntertainmentCredits> =
         when (movieCreditsResult) {
-            is Result.Success -> Result.Success(entertainmentMapper.createUIEntertainmentCredits(movieCreditsResult.data))
+            is Result.Success -> Result.Success(coreMapper.createUIEntertainmentCredits(movieCreditsResult.data))
             else -> mapErrorOrException(movieCreditsResult)
         }
 
@@ -43,7 +42,7 @@ class MovieDetailsMapper(private val entertainmentMapper: IEntertainmentMapper) 
             movieDetails.release_date,
             movieDetails.vote_count,
             movieDetails.vote_average,
-            movieDetails.genres.map { createUIGenre(it) }
+            movieDetails.genres.map { coreMapper.createUIGenre(it) }
         )
 
     private fun createUITrailerList(trailer: NetworkTrailersList) =
@@ -53,6 +52,4 @@ class MovieDetailsMapper(private val entertainmentMapper: IEntertainmentMapper) 
         UITrailer(trailer.site, trailer.key)
 
     private fun createUICountry(item: NetworkObjectWithName) = UICountry(item.name)
-
-    private fun createUIGenre(item: NetworkObjectWithName) = UIGenre(item.name)
 }
