+++ app/src/main/java/com/kirchhoff/movies/mapper/tv/TvDetailsMapper.kt
@@ -1,16 +1,14 @@
 package com.kirchhoff.movies.mapper.tv
 
 import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.data.network.core.NetworkObjectWithName
 import com.kirchhoff.movies.data.network.details.tv.NetworkTvDetails
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
-import com.kirchhoff.movies.data.ui.core.UIGenre
 import com.kirchhoff.movies.data.ui.details.tv.UITvDetails
 import com.kirchhoff.movies.mapper.BaseMapper
-import com.kirchhoff.movies.mapper.entertainment.IEntertainmentMapper
+import com.kirchhoff.movies.mapper.core.ICoreMapper
 import com.kirchhoff.movies.repository.Result
 
-class TvDetailsMapper(private val entertainmentMapper: IEntertainmentMapper) : BaseMapper(), ITvDetailsMapper {
+class TvDetailsMapper(private val coreMapper: ICoreMapper) : BaseMapper(), ITvDetailsMapper {
 
     override fun createUITvDetails(tvDetailsResult: Result<NetworkTvDetails>): Result<UITvDetails> =
         when (tvDetailsResult) {
@@ -20,7 +18,7 @@ class TvDetailsMapper(private val entertainmentMapper: IEntertainmentMapper) : B
 
     override fun createUIEntertainmentCredits(tvCreditsResult: Result<NetworkEntertainmentCredits>): Result<UIEntertainmentCredits> =
         when (tvCreditsResult) {
-            is Result.Success -> Result.Success(entertainmentMapper.createUIEntertainmentCredits(tvCreditsResult.data))
+            is Result.Success -> Result.Success(coreMapper.createUIEntertainmentCredits(tvCreditsResult.data))
             else -> mapErrorOrException(tvCreditsResult)
         }
 
@@ -33,9 +31,6 @@ class TvDetailsMapper(private val entertainmentMapper: IEntertainmentMapper) : B
             tvDetails.first_air_date,
             tvDetails.vote_count,
             tvDetails.vote_average,
-            tvDetails.genres.map { createUIGenre(it) }
+            tvDetails.genres.map { coreMapper.createUIGenre(it) }
         )
-
-    private fun createUIGenre(item: NetworkObjectWithName) =
-        UIGenre(item.name)
 }
